remove the hardcoded paths in count_lines.sh

+++ verus-mimalloc/count_lines.sh
@@ -1,12 +1,18 @@
-DIR=/Users/tjhance/Dropbox/verus-systems-code/memory-allocators/verus-mimalloc
-LCDIR=/Users/tjhance/Dropbox/verus/source/tools/line_count
+DIR=$(dirname "$0")
+
+if [ ! -d "$LINE_COUNT_DIR" ]; then
+  echo "set \$LINE_COUNT_DIR to the directory verus/source/tools/line_count"
+  exit -1
+fi
 
 rm -rf lib.d
 ./rrv --emit=dep-info --no-verify
 
-cd $LCDIR
+pushd $LINE_COUNT_DIR
 cargo run --release -- $DIR/lib.d -p 2> ~/o
 cargo run --release -- $DIR/lib.d --json > $DIR/lc.json
+popd
 
-cd $DIR
+pushd $DIR
 python3 process_lines.py < lc.json
+popd
